requirements.txt
huggingface_hub==0.25.0
python-dotenv==1.0.1
jsonschema==4.22.0
